+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 ## Release Notes 3.12.0
 
 ## Changes
+
+- added tournaments section
+- added friends
  
 ## Bugfixes